Update the addBooks method, with tests
I have updated the function of checking whether the bookId is duplicated in the database of the addBooks method in the class of booksOrganization. And added a few lines of test code, and the test code can run successfully. And the connection between Java and the database is working without any problem.